Rename formik related content components
